requirements.txt
absl-py==0.9.0
astor==0.8.1
atari-py==0.2.6
Click==7.0
cloudpickle==1.2.2
decorator==4.4.1
dict-recursive-update==1.0.1
future==0.18.2
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.2
gym==0.15.7
h5py==2.10.0
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
mock==4.0.1
networkx==2.4
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
protobuf==3.11.3
pyglet==1.5.0
PyYAML==5.3
scipy==1.4.1
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.42.1
Werkzeug==1.0.0
wrapt==1.12.0
